+++ src/g_game.c
@@ -1132,9 +1132,7 @@ void G_PlayerReborn(void)
 //
 static void G_DoReborn(void)
 {
-    if (solonet)
-        P_ResurrectPlayer(initial_health);
-    else if (quicksaveslot >= 0 && autoload && !pistolstart)
+    if (quicksaveslot >= 0 && autoload && !pistolstart)
         gameaction = ga_autoloadgame;
     else
     {