+++ src/p_user.c
@@ -322,7 +322,7 @@ static void P_DeathThink(void)
     weaponrumbletics = 1;
     idlechainsawrumblestrength = 0;
     freeze = false;
-    infight = (infighting && !solonet && !(viewplayer->cheats & CF_NOTARGET));
+    infight = (infighting && !(viewplayer->cheats & CF_NOTARGET));
 
     P_MovePlayerSprites();